capitulo_3/recursos/ejercicio_3/api/requirements.txt
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.4.0
Flask==1.1.1
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-bigquery==1.22.0
google-cloud-core==1.0.3
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==7.2.0
openapi-spec-validator==0.2.8
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyrsistent==0.15.5
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
six==1.13.0
swagger-ui-bundle==0.0.6
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
